C.R.M.-Project-main/mysite/requirements-dev.txt
# Development Dependencies

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
coverage==7.3.2
faker==20.1.0

# Code Quality
flake8==6.1.0
black==23.11.0
isort==5.12.0
pylint==3.0.2
autopep8==2.0.4

# Security
bandit==1.7.5
safety==2.3.5

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Debugging
django-debug-toolbar==4.2.0
ipython==8.17.2
ipdb==0.13.13

# Performance
django-silk==5.0.4

# Type Checking
mypy==1.7.1
django-stubs==4.2.6
